chore: move fitMode to flamegraph
